+++ packages/tracing-internal/src/browser/backgroundtab.ts
