+++ packages/tracing-internal/src/browser/browsertracing.ts
