+++ packages/tracing-internal/src/browser/index.ts
