+++ packages/tracing-internal/src/browser/metrics/utils.ts
