+++ packages/tracing-internal/src/browser/router.ts
