+++ packages/tracing-internal/src/browser/types.ts
