+++ packages/tracing-internal/src/browser/web-vitals/getFID.ts
