+++ packages/tracing-internal/src/browser/web-vitals/getLCP.ts
