+++ packages/tracing-internal/src/browser/web-vitals/lib/bindReporter.ts
