+++ packages/tracing-internal/src/browser/web-vitals/lib/getActivationStart.ts
