+++ packages/tracing-internal/src/browser/web-vitals/lib/getNavigationEntry.ts
