+++ packages/tracing-internal/src/browser/web-vitals/lib/getVisibilityWatcher.ts
