+++ packages/tracing-internal/src/browser/web-vitals/lib/initMetric.ts
