+++ packages/tracing-internal/src/browser/web-vitals/lib/observe.ts
