+++ packages/tracing-internal/src/browser/web-vitals/lib/onHidden.ts
