+++ packages/tracing-internal/src/browser/web-vitals/types.ts
