+++ packages/tracing-internal/src/browser/web-vitals/types/base.ts
