+++ packages/tracing-internal/src/browser/web-vitals/types/cls.ts
