+++ packages/tracing-internal/src/browser/web-vitals/types/lcp.ts
