+++ packages/tracing-internal/src/browser/web-vitals/types/polyfills.ts
