+++ packages/tracing-internal/src/errors.ts
