+++ packages/tracing-internal/src/extensions.ts
