+++ packages/tracing-internal/src/node/integrations/graphql.ts
