+++ packages/tracing/test/integrations/graphql.test.ts
@@ -2,8 +2,7 @@
 import { Hub, Scope } from '@sentry/core';
 import { logger } from '@sentry/utils';
 
-import { Span } from '../../src';
-import { GraphQL } from '../../src/node/integrations/graphql';
+import { Integrations, Span } from '../../src';
 import { getTestClient } from '../testutils';
 
 const GQLExecute = {
@@ -33,7 +32,7 @@ describe('setupOnce', () => {
   let childSpan: Span;
 
   beforeAll(() => {
-    new GraphQL().setupOnce(
+    new Integrations.GraphQL().setupOnce(
       () => undefined,
       () => new Hub(undefined, scope),
     );
@@ -66,7 +65,7 @@ describe('setupOnce', () => {
     const client = getTestClient({ instrumenter: 'otel' });
     const hub = new Hub(client);
 
-    const integration = new GraphQL();
+    const integration = new Integrations.GraphQL();
     integration.setupOnce(
       () => {},
       () => hub,
